Fix task 0

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -6,7 +6,7 @@
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	if (!h)
 		return (count);
@@ -14,7 +14,7 @@ size_t print_dlistint(const dlistint_t *h)
 		h = h->prev;
 	while (h != NULL)
 	{
-		printf("%u\n", h->n);
+		printf("%d\n", h->n);
 		count += 1;
 		h = h->next;
 	}
